make NodeActionsGroup
