+++ apps/src/sketchlab/reactFlow/elementToolbars/ImageNodeToolbar.tsx
@@ -1,13 +1,11 @@
-import {useReactFlow} from '@xyflow/react';
 import React from 'react';
 
 import {DEFAULT_ROTATION} from '../constants';
 import {ImageNodeType} from '../types';
-import {nextBackZIndex, nextFrontZIndex} from '../utils/stacking';
 
-import ActionsGroup from './ActionsGroup';
 import HandleVisibilityToggle from './HandleVisibilityToggle';
 import LockedNotice from './LockedNotice';
+import NodeActionsGroup from './NodeActionsGroup';
 import RotationGroup from './RotationGroup';
 import ToolbarShell from './ToolbarShell';
 import {useNodeToolbarData} from './useNodeToolbarData';
@@ -17,7 +15,6 @@ interface ImageNodeToolbarProps {
 }
 
 export default function ImageNodeToolbar({nodeId}: ImageNodeToolbarProps) {
-  const {deleteElements, updateNode, getNodes, getEdges} = useReactFlow();
   const {data, patchNodeData} = useNodeToolbarData<ImageNodeType>(nodeId);
   const handlesVisible = data.showHandles !== false;
 
@@ -35,18 +32,7 @@ export default function ImageNodeToolbar({nodeId}: ImageNodeToolbarProps) {
             value={data.rotation ?? DEFAULT_ROTATION}
             onChange={degrees => patchNodeData({rotation: degrees})}
           />
-          <ActionsGroup
-            onDelete={() => deleteElements({nodes: [{id: nodeId}]})}
-            onLock={() => patchNodeData({locked: true})}
-            onBringToFront={() => {
-              const items = [...getNodes(), ...getEdges()];
-              updateNode(nodeId, {zIndex: nextFrontZIndex(items, nodeId)});
-            }}
-            onSendToBack={() => {
-              const items = [...getNodes(), ...getEdges()];
-              updateNode(nodeId, {zIndex: nextBackZIndex(items, nodeId)});
-            }}
-          />
+          <NodeActionsGroup nodeId={nodeId} />
           <HandleVisibilityToggle
             visible={handlesVisible}
             onToggle={() => patchNodeData({showHandles: !handlesVisible})}
