+++ apps/src/sketchlab/reactFlow/elementToolbars/NodeActionsGroup.tsx
@@ -1,0 +1,29 @@
+import {useReactFlow} from '@xyflow/react';
+import React from 'react';
+
+import {nextBackZIndex, nextFrontZIndex} from '../utils/stacking';
+
+import ActionsGroup from './ActionsGroup';
+
+interface NodeActionsGroupProps {
+  nodeId: string;
+}
+
+export default function NodeActionsGroup({nodeId}: NodeActionsGroupProps) {
+  const {deleteElements, updateNode, updateNodeData, getNodes, getEdges} =
+    useReactFlow();
+  return (
+    <ActionsGroup
+      onDelete={() => deleteElements({nodes: [{id: nodeId}]})}
+      onLock={() => updateNodeData(nodeId, {locked: true})}
+      onBringToFront={() => {
+        const items = [...getNodes(), ...getEdges()];
+        updateNode(nodeId, {zIndex: nextFrontZIndex(items, nodeId)});
+      }}
+      onSendToBack={() => {
+        const items = [...getNodes(), ...getEdges()];
+        updateNode(nodeId, {zIndex: nextBackZIndex(items, nodeId)});
+      }}
+    />
+  );
+}
