+++ apps/src/sketchlab/reactFlow/elementToolbars/ShapeNodeToolbar.tsx
@@ -1,14 +1,12 @@
-import {useReactFlow} from '@xyflow/react';
 import React from 'react';
 
 import {DEFAULT_ROTATION} from '../constants';
 import {ShapeNodeType} from '../types';
-import {nextBackZIndex, nextFrontZIndex} from '../utils/stacking';
 
-import ActionsGroup from './ActionsGroup';
 import FontSizeGroup from './FontSizeGroup';
 import HandleVisibilityToggle from './HandleVisibilityToggle';
 import LockedNotice from './LockedNotice';
+import NodeActionsGroup from './NodeActionsGroup';
 import RotationGroup from './RotationGroup';
 import SwatchGroup from './SwatchGroup';
 import {
@@ -28,7 +26,6 @@ interface ShapeNodeToolbarProps {
 }
 
 export default function ShapeNodeToolbar({nodeId}: ShapeNodeToolbarProps) {
-  const {deleteElements, updateNode, getNodes, getEdges} = useReactFlow();
   const {data, patchNodeData} = useNodeToolbarData<ShapeNodeType>(nodeId);
 
   const {backgroundColor, strokeColor, fontSize, fontColor} = data;
@@ -72,18 +69,7 @@ export default function ShapeNodeToolbar({nodeId}: ShapeNodeToolbarProps) {
             value={data.rotation ?? DEFAULT_ROTATION}
             onChange={degrees => patchNodeData({rotation: degrees})}
           />
-          <ActionsGroup
-            onDelete={() => deleteElements({nodes: [{id: nodeId}]})}
-            onLock={() => patchNodeData({locked: true})}
-            onBringToFront={() => {
-              const items = [...getNodes(), ...getEdges()];
-              updateNode(nodeId, {zIndex: nextFrontZIndex(items, nodeId)});
-            }}
-            onSendToBack={() => {
-              const items = [...getNodes(), ...getEdges()];
-              updateNode(nodeId, {zIndex: nextBackZIndex(items, nodeId)});
-            }}
-          />
+          <NodeActionsGroup nodeId={nodeId} />
           <HandleVisibilityToggle
             visible={handlesVisible}
             onToggle={() => patchNodeData({showHandles: !handlesVisible})}
